OSIS-3859 Add tests + refactoring
